feat: 🎸 migrate toolbar to vue
