+++ packages/components/package.json
@@ -106,7 +106,6 @@
     "@codemirror/view": "^6"
   },
   "dependencies": {
-    "@atomico/hooks": "^4.1.2",
     "@floating-ui/dom": "^1.5.1",
     "@milkdown/core": "workspace:*",
     "@milkdown/ctx": "workspace:*",
@@ -119,7 +118,6 @@
     "@milkdown/utils": "workspace:*",
     "@types/lodash.debounce": "^4.0.7",
     "@types/lodash.throttle": "^4.1.9",
-    "atomico": "^1.75.1",
     "clsx": "^2.0.0",
     "dompurify": "^3.2.5",
     "lodash.debounce": "^4.0.8",
