+++ packages/components/src/index.ts
@@ -1,2 +1,1 @@
-export { html } from 'atomico'
 export { Icon } from './__internal__/components/icon'
