+++ packages/crepe/package.json
@@ -63,7 +63,6 @@
     "@codemirror/theme-one-dark": "^6.1.2",
     "@codemirror/view": "^6.16.0",
     "@milkdown/kit": "workspace:*",
-    "atomico": "^1.75.1",
     "clsx": "^2.0.0",
     "codemirror": "^6.0.1",
     "katex": "^0.16.0",
