+++ packages/crepe/src/icons/bold.ts
@@ -1,6 +1,4 @@
-import { html } from 'atomico'
-
-export const boldIcon = html`
+export const boldIcon = `
   <svg
     xmlns="http://www.w3.org/2000/svg"
     width="24"
