+++ packages/crepe/src/icons/code.ts
@@ -1,25 +1,3 @@
-import { html } from 'atomico'
-
-export const legacyCodeIcon = html`
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-  >
-    <g clip-path="url(#clip0_977_8081)">
-      <path
-        d="M9.4 16.6L4.8 12L9.4 7.4L8 6L2 12L8 18L9.4 16.6ZM14.6 16.6L19.2 12L14.6 7.4L16 6L22 12L16 18L14.6 16.6Z"
-      />
-    </g>
-    <defs>
-      <clipPath id="clip0_977_8081">
-        <rect width="24" height="24" />
-      </clipPath>
-    </defs>
-  </svg>
-`
-
 export const codeIcon = `
   <svg
     xmlns="http://www.w3.org/2000/svg"
