+++ packages/crepe/src/icons/functions.ts
@@ -1,19 +1,3 @@
-import { html } from 'atomico'
-
-export const legacyFunctionsIcon = html`
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="32"
-    height="32"
-    viewBox="0 0 24 24"
-  >
-    <path
-      fill="currentColor"
-      d="M7 19v-.808L13.096 12L7 5.808V5h10v1.25H9.102L14.727 12l-5.625 5.77H17V19z"
-    />
-  </svg>
-`
-
 export const functionsIcon = `
   <svg
     xmlns="http://www.w3.org/2000/svg"
