+++ packages/crepe/src/icons/italic.ts
@@ -1,6 +1,4 @@
-import { html } from 'atomico'
-
-export const italicIcon = html`
+export const italicIcon = `
   <svg
     xmlns="http://www.w3.org/2000/svg"
     width="24"
