+++ packages/crepe/src/icons/strikethrough.ts
@@ -1,6 +1,4 @@
-import { html } from 'atomico'
-
-export const strikethroughIcon = html`
+export const strikethroughIcon = `
   <svg
     xmlns="http://www.w3.org/2000/svg"
     width="24"
