+++ storybook/package.json
@@ -20,7 +20,6 @@
     "@milkdown/theme-nord": "workspace:*",
     "@uiw/codemirror-theme-basic": "^4.22.1",
     "@uiw/codemirror-theme-nord": "^4.22.1",
-    "atomico": "^1.75.1",
     "codemirror": "^6.0.1",
     "katex": "^0.16.0",
     "shiki": "^3.0.0",
